requirements.txt
asgiref==3.8.1
Django==4.2.7
djangorestframework==3.15.1
psycopg2==2.9.9
sqlparse==0.4.4
tzdata==2024.1
